final commit

+++ src/main/java/com/assignment/notificationservice/controller/redis/BlacklistedNumbersController.java
@@ -3,17 +3,13 @@ package com.assignment.notificationservice.controller.redis;
 import com.assignment.notificationservice.dao.redis.BlacklistedNumbersRepository;
 import com.assignment.notificationservice.dto.requestDTO.redis.BlacklistedNumbersDTO;
 import com.assignment.notificationservice.dto.responseDTO.redis.BlacklistedNumbersResponseDTO;
-import com.assignment.notificationservice.dto.responseDTO.api.GeneralMessageDTO;
 import com.assignment.notificationservice.entity.redis.BlacklistedNumbers;
-import com.assignment.notificationservice.exception.InternalServerErrorException;
+import com.assignment.notificationservice.exception.BadRequestErrorException;
 import com.assignment.notificationservice.exception.RedisInteralServerErrorException;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -36,15 +32,22 @@ public class BlacklistedNumbersController {
     @Value("${REDIS_KEY}")
     private String KEY;
 
+    public Boolean isValidNumbers(List<String> phoneNumbers){
+        for (String str : phoneNumbers) {
+            if (!str.matches("[0-9]+") || str.length() < 7 || str.length() > 10) {
+                return false;
+            }
+        }
+        return true;
+    }
     @GetMapping("/")
     public ResponseEntity<BlacklistedNumbersResponseDTO> findAllBlacklistedNumbers() throws RedisInteralServerErrorException {
-        log.debug("Retrieving all blacklisted numbers");
         try {
             List<String> result = redisTemplate.opsForHash().values(KEY);
             log.debug("Retrieved {} blacklisted numbers successfully", result.size());
             return ResponseEntity.ok(new BlacklistedNumbersResponseDTO(result));
         } catch (Exception e) {
-            log.error("Error occurred while retrieving blacklisted numbers from redis", e);
+            log.error("Error occurred while retrieving blacklisted numbers from redis");
            throw new RedisInteralServerErrorException("Error occurred while retrieving blacklisted numbers");
         }
     }
@@ -52,16 +55,16 @@ public class BlacklistedNumbersController {
     @PostMapping("/")
     public ResponseEntity<BlacklistedNumbersResponseDTO> addNewNumbersToBlacklistedNumbers(@RequestBody BlacklistedNumbersDTO blacklistedNumbersDTO) throws RedisInteralServerErrorException {
         List<String> phoneNumbers = blacklistedNumbersDTO.getPhoneNumbers();
-        log.debug(String.format("The list of numbers recieved to be blacklisted : %s", phoneNumbers.toString()));
+        if(!isValidNumbers(phoneNumbers)){
+            throw new BadRequestErrorException();
+        }
         List<String> successfullyBlacklistedNumbers=new ArrayList<>();
         for (String phoneNumber : phoneNumbers) {
             try {
                 if (redisTemplate.opsForHash().hasKey(KEY, phoneNumber)) {
-                    log.debug("Phone number %s is already blocked.", phoneNumber);
                     continue;
                 }
                 redisTemplate.opsForHash().put(KEY, phoneNumber, phoneNumber);
-                log.debug("Added phone number %s to redis blacklisted list.", phoneNumber);
                 var blacklistedNumber = new BlacklistedNumbers(phoneNumber);
                 blacklistedNumbersRepository.save(blacklistedNumber);
                 log.debug("Added phone number %s to blacklisted database.", phoneNumber);
@@ -76,15 +79,16 @@ public class BlacklistedNumbersController {
     @DeleteMapping("/")
     public ResponseEntity<BlacklistedNumbersResponseDTO> deleteNumbersFromBlacklistedNumbers(@RequestBody BlacklistedNumbersDTO blacklistedNumbersDTO) throws RedisInteralServerErrorException {
         List<String> phoneNumbers = blacklistedNumbersDTO.getPhoneNumbers();
+        if(!isValidNumbers(phoneNumbers)){
+            throw new BadRequestErrorException();
+        }
         List<String> successfullyBlacklistedNumbers = new ArrayList<>();
         for (String phoneNumber : phoneNumbers) {
             try {
                 if (!redisTemplate.opsForHash().hasKey(KEY, phoneNumber)) {
-                    log.debug("Phone number {} doesn't exist in the database.",phoneNumber);
                     continue;
                 }
                 redisTemplate.opsForHash().delete(KEY, phoneNumber);
-                log.debug(String.format("Deleted phone number %s from redis.", phoneNumber));
                 blacklistedNumbersRepository.deleteSpecificPhoneNumber(phoneNumber);
                 log.debug(String.format("Deleted phone number %s from databse.", phoneNumber));
                 successfullyBlacklistedNumbers.add(phoneNumber);
